- Unified Allocation pattern As previously mentioned, the allocation pattern was a bit inconsistent where the modules doesn't require to be deinit, but handled by the project load and slot deletion. Thus, I have finalized the type with the following pattern:
If the function doesn't return a collection, this
suggests that the heap allocation are done internally
that depends on the Slot type, so users won't need
to specify an allocator because the deletion will
be done by the slot type.

If the function returns a non string slices, it
is expected for users to manually manage the memory
of the return slice, or specify an input slice
during function call because these type don't closely
related to the slot.