- Cross Compilation
This commit address the lack of other Compilation
targets, including Mac and x86/x64 version of
Linux and windows, which should covers most of the
common operating systems and CPU architectures.
Arm Linux will be added in the future.

The next function will be worked on is to provide
a better parameter and return type for zig, and
trying to get the audio callback working such that
it is possible to use SunVox as an extension,
fetching the audio data for other processes.

For the async nature to the controller change, there
will be a queue to handle controller event change
in bulk, and to ensure the integrity before saving
the project, but this will require the future version
of zig (possibly 0.16.0) for the async features.